fix(ci): maintainability line-limit — split compose env interpolation into compose_env
src/awf/profiles/compose.py grew to 1659 lines, exceeding the 1500-line
first-party file limit enforced by
test_first_party_code_files_stay_under_line_limit (PR #751 CI failure).

Extract the Compose env-value interpolation / resolution machinery
(_ComposeEnvResolution enum, _COMPOSE_* constants, and the pure resolve/
concrete-password/expanded-value/environment-mapping helpers) into a new
src/awf/profiles/compose_env.py module (486 lines). compose.py re-imports
every moved name so all existing callers and tests — including
from awf.profiles.compose import _COMPOSE_PASSTHROUGH and
compose_module.<name> attribute access — keep working unchanged. This is a
pure relocation with no behavior change; compose.py is now 1218 lines.
